backend/requirements.txt
async-timeout==4.0.2
bcrypt==4.0.1
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==1.4.2
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.5
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
